memory errors are fixed
git-svn-id: https://nocs.stanford.edu/cgi-bin/svn.cgi/booksim2.0/branches/hotspot_reservation@4119 ce7aba0d-9f81-40a6-a495-cecae84b0515

+++ src/flit.hpp
@@ -50,7 +50,7 @@ public:
   int exptime;
   int  sn;
   int flid;
-  short payload;
+  int payload;
   int head_sn;
 
   int flbid;
